+++ src/main/java/com/ejoongseok/wmslive/outbound/domain/Inventories.java
@@ -1,4 +1,4 @@
-package com.ejoongseok.wmslive.outbound.feature;
+package com.ejoongseok.wmslive.outbound.domain;
 
 import com.ejoongseok.wmslive.location.domain.Inventory;
 
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public record Inventories(List<Inventory> inventories, Long orderQuantity) {
-    void validateInventory() {
+    public void validateInventory() {
         final long totalInventoryQuantity = inventories().stream()
                 .filter(i -> 0L < i.getInventoryQuantity())
                 .filter(i -> i.getLpn().getExpirationAt().isAfter(LocalDateTime.now()))
