+++ src/main/java/com/ejoongseok/wmslive/outbound/domain/PackagingMaterials.java
@@ -1,13 +1,11 @@
-package com.ejoongseok.wmslive.outbound.feature;
-
-import com.ejoongseok.wmslive.outbound.domain.PackagingMaterial;
+package com.ejoongseok.wmslive.outbound.domain;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 
 public record PackagingMaterials(List<PackagingMaterial> packagingMaterials) {
-    Optional<PackagingMaterial> findOptimalPackagingMaterial(
+    public Optional<PackagingMaterial> findOptimalPackagingMaterial(
             final Long totalWeight,
             final Long totalVolume) {
         return packagingMaterials().stream()
